// packages/oas-utils/src/migrations/v-2.1.0/types.generated.ts
export namespace v_2_1_0 {
// ...
  export type RequestExample = {
// ...
    parameters?: {
      path?: {
        key?: string
        value?: string
        enabled?: boolean
        file?: any | undefined
        description?: string | undefined
        refUid?: string | undefined
        required?: boolean | undefined
        enum?: string[] | undefined
        type?: string | undefined
        format?: string | undefined
        minimum?: number | undefined
        maximum?: number | undefined
        default?: any | undefined
        nullable?: boolean | undefined
      }[]
      query?: {
        key?: string
        value?: string
        enabled?: boolean
        file?: any | undefined
        description?: string | undefined
        refUid?: string | undefined
        required?: boolean | undefined
        enum?: string[] | undefined
        type?: string | undefined
        format?: string | undefined
        minimum?: number | undefined
        maximum?: number | undefined
        default?: any | undefined
        nullable?: boolean | undefined
      }[]
      headers?: {
        key?: string
        value?: string
        enabled?: boolean
        file?: any | undefined
        description?: string | undefined
        refUid?: string | undefined
        required?: boolean | undefined
        enum?: string[] | undefined
        type?: string | undefined
        format?: string | undefined
        minimum?: number | undefined
        maximum?: number | undefined
        default?: any | undefined
        nullable?: boolean | undefined
      }[]
      cookies?: {
        key?: string
        value?: string
        enabled?: boolean
        file?: any | undefined
        description?: string | undefined
        refUid?: string | undefined
        required?: boolean | undefined
        enum?: string[] | undefined
        type?: string | undefined
        format?: string | undefined
        minimum?: number | undefined
        maximum?: number | undefined
        default?: any | undefined
        nullable?: boolean | undefined
      }[]
    }
  }

  export type Request = {
// ...
    method?:
      | 'connect'
      | 'delete'
      | 'get'
      | 'head'
      | 'options'
      | 'patch'
      | 'post'
      | 'put'
      | 'trace'
    servers?: string[]
    selectedServerUid?: string
    examples?: string[]
    selectedSecuritySchemeUids?: string[]
  }
// ...
}